refactor: 계층 분리
