+++ src/main/java/attendance/domain/AttendanceStatus.java
@@ -1,4 +1,4 @@
-package attendance;
+package attendance.domain;
 
 public enum AttendanceStatus {
     LATE,
