+++ src/main/java/attendance/domain/AttendanceTime.java
@@ -1,4 +1,4 @@
-package attendance;
+package attendance.domain;
 
 import java.time.LocalTime;
