+++ src/main/java/attendance/domain/Warning.java
@@ -1,4 +1,4 @@
-package attendance;
+package attendance.domain;
 
 public enum Warning {
     Warning,
