+++ src/test/java/attendance/domain/AttendanceTimeTest.java
@@ -1,4 +1,4 @@
-package attendance;
+package attendance.domain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
